DONE2

+++ AQPHashTable.java
@@ -22,16 +22,16 @@ public class AQPHashTable extends OAHashTable {
 
 	@Override
 	public int Hash(long x, int i) {
-		int result;
+		long result;
 		if (i%2 == 0){ //if (-1)^i==1
-			result= (int) (this.modHash.Hash(x)+Math.pow(i,2))%m;
+			result= (long) (this.modHash.Hash(x)+Math.pow(i,2))%m;
 		}
 		else { //if (-1)^i==-1
-			result = (int) (this.modHash.Hash(x) - Math.pow(i, 2)) % m;
+			result = (long) (this.modHash.Hash(x) - Math.pow(i, 2)) % m;
 
 			if (result < 0) //this line makes sure result>=0
 				result += m;
 		}
-		return result;
+		return (int) result;
 	}
 }
